refactor: ChatSection 채팅 로직을 useBattleChat 훅으로 분리
- useBattleChat 훅 생성 후 기존 ChatSection에서 있던  채팅 메시지 필터링/변환 및 socket 이벤트 구독 로직 이번 훅으로 이동
- chatSection에서 채팅 로그 삭제후 store에서 chats 상태 구독하도록 로직 변경
- Message 타입 및 변환 함수를 utils/convertChatMessage.ts로 분리
